[backend] feat: 일기 작성 API v2 컨트롤러 및 서비스 계층 구현

+++ backend/src/main/java/com/example/demo/dto/DiaryWriteRequestDTOv2.java
@@ -11,5 +11,5 @@ public class DiaryWriteRequestDTOv2 {
     private String writtenDate;
     private String title;
     private String details;
-    private List<Integer> sharedTeamList = new ArrayList<>();
+    private List<Long> sharedTeamList = new ArrayList<>();
 }
